Fixed broken file transfer introduced by aramid
scp and cf-remote default to transferring files to the home folder,
where you end up if you ssh to the machine. aramid defaults to
using the same path as in src, which might not exist or be accessible.
Restored the previous behavior by explicitly giving destination
path to aramid library.

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>