Fixed print/debug messages when running commands on localhost
Fixed some earlier here: a77369aedb52a6869fe45b6cda3195e5ef8f2839
but missed a few.

The underlying issue is that aramid result.stdout and stderr
are strings but when we connect to localhost we use a subprocess
and result.stdout/stderr are byte sequences and so can't be appended
to each other as strings can.

Ticket: CFE-4024
Changelog: title