CFE-4322: Changed how cf-remote spawn finds AMIs with AWS
Before we hard-coded AMI and had to update them fairly often and add them when we supported a new platform.

Here we change to querying for AMI based on known trusted owner IDs.

This should make it so that new platforms are automatically available and any updates will be in-place automatically as well.

This change should also enable users to specify other regions and have most tplatforms "just work".

One note: specific Windows AMI are only available in eu-west-1

Ticket: CFE-4322
Changelog: title

Co-authored-by: Lars Erik Wik <lars.erik.wik@northern.tech>