Improved handling of unexpected exceptions
In my case, I was encountering another exception, FileNotFoundError.
This is not expected, and it makes sense to use the nicer printing
instead of a big ugly stacktrace. In many cases the single line
error message is enough, and the stacktrace doesn't add any value.
You can still set the env var to get the full stacktrace.

Signed-off-by: Ole Herman Schumacher Elgesem <ole.elgesem@northern.tech>